+++ simulator/compiler/parser/grammar/rdobison.h
@@ -12,6 +12,7 @@
 #define _RDOBISON_H_
 
 // ----------------------------------------------------------------------- INCLUDES
+#include <cstddef>
 // ----------------------------------------------------------------------- SYNOPSIS
 // --------------------------------------------------------------------------------
